QA | IdnaEncoder::punycode_encode(): pass the correct type
The `IdnaEncoder::digit_to_char()` method expects an `int`, not a float, for the `$digit` parameter. Same goes for the PHP native `substr()` function called from within the `IdnaEncoder::digit_to_char()` method to which the `$digit` parameter.

This ensures that the values passed to `IdnaEncoder::digit_to_char()`, which due to the use of `%` and `floor` are typed as `float`, will now contain the correct `int` type.

Note: no tests are added for this as the existing tests cover this already, but don't show the issue unless `strict_types` mode is turned on.

+++ src/IdnaEncoder.php
@@ -329,10 +329,10 @@ class IdnaEncoder {
 						}
 
 						// output the code point for digit t + ((q - t) mod (base - t))
-						$digit   = $t + (($q - $t) % (self::BOOTSTRAP_BASE - $t));
+						$digit   = (int) ($t + (($q - $t) % (self::BOOTSTRAP_BASE - $t)));
 						$output .= self::digit_to_char($digit);
 						// let q = (q - t) div (base - t)
-						$q = floor(($q - $t) / (self::BOOTSTRAP_BASE - $t));
+						$q = (int) floor(($q - $t) / (self::BOOTSTRAP_BASE - $t));
 					} // end
 					// output the code point for digit q
 					$output .= self::digit_to_char($q);
